+++ examples/web/src/main.rs
@@ -1,11 +1,12 @@
 pub mod mailer;
 pub mod shared;
+pub mod sse;
 pub mod users;
 
 use dotenv::dotenv;
 use sword::prelude::*;
 
-use crate::{mailer::MailerModule, shared::SharedModule, users::UsersModule};
+use crate::{mailer::MailerModule, shared::SharedModule, sse::SseModule, users::UsersModule};
 
 #[sword::main]
 async fn main() {
@@ -17,6 +18,7 @@ async fn main() {
         .with_module::<SharedModule>()
         .with_module::<UsersModule>()
         .with_module::<MailerModule>()
+        .with_module::<SseModule>()
         .build();
 
     app.run().await;