+++ examples/web/src/sse/mod.rs
@@ -1,0 +1,13 @@
+mod controller;
+
+use controller::SseController;
+
+use sword::prelude::*;
+
+pub struct SseModule;
+
+impl Module for SseModule {
+    fn register_controllers(controllers: &ControllerRegistry) {
+        controllers.register::<SseController>();
+    }
+}